+++ CHANGELOG.md
@@ -7,6 +7,7 @@ _none at the moment_
 ###2.5.9 (01-16-2015)
 * Removed progressing loader when all slides are video
 * Improved fading out progressing loader
+* Associated MSMGT to SMGT logo in the notes area when there are not notes
 * Updated kWidget to version 2.25
 * Google Chrome resolved the HTML5 video text aliasing issue on Windows, so HTML5 video player is back.